requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-contrib-python==4.9.0.80
numpy==1.26.3
cupy-cuda12x==12.3.0
pydantic-settings==2.1.0
aiohttp==3.9.1
uvloop==0.19.0
